move file/redis stores into separate files to encapsulate implementation
